requirements.txt
matplotlib==3.5.2
numpy==1.21.6
pandas==1.3.5
Pillow==9.1.1
scikit_learn==1.1.1
streamlit==1.9.0
streamlit_option_menu==0.3.2
